Adding random data to json to test tooltip
Moved tooltip to left of pointer so we can always see the contents even
when hovering over the Shetland Isles.
Having our data in the json makes looking it up in the tooltips very
easy.  However, it's going to be a very tedious process to convert our
xls data into the correct json format and integrate it with the map
data.